+++ modules/farm_rothamsted_experiment/farm_rothamsted_experiment.post_update.php
@@ -871,3 +871,12 @@ function farm_rothamsted_experiment_post_update_2_13_uninstall_contact_field(&$s
   $contact_field = $update_manager->getFieldStorageDefinition('contact', 'plan');
   \Drupal::entityDefinitionUpdateManager()->uninstallFieldStorageDefinition($contact_field);
 }
+
+/**
+ * Uninstall experiment_code field.
+ */
+function farm_rothamsted_experiment_post_update_2_14_uninstall_experiment_code_field(&$sandbox = NULL) {
+  $update_manager = \Drupal::entityDefinitionUpdateManager();
+  $contact_field = $update_manager->getFieldStorageDefinition('experiment_code', 'plan');
+  \Drupal::entityDefinitionUpdateManager()->uninstallFieldStorageDefinition($contact_field);
+}